[audioCoreDevice] Update to newer audioDevice api, patch by Nibbles

+++ branches/avidemux_2.5_branch_gruntster/plugins/ADM_audioDevices/AudioCore/ADM_deviceAudioCore.h
@@ -17,7 +17,7 @@
 					uint8_t				_inUse;
 		  public:
 		  				coreAudioDevice(void);
-		     		virtual uint8_t init(uint8_t channels, uint32_t fq);
+		     		virtual uint8_t init(uint32_t channels, uint32_t fq);
 	    			virtual uint8_t play(uint32_t len, float *data);
 		      		virtual uint8_t stop(void);
 					virtual uint8_t setVolume(int volume);
